LayoutTests/editing/async-clipboard/clipboard-read-text-expected.txt

This test verifies that navigator.clipboard.readText can be used to read plain text from the system clipboard. To verify manually, click the 'Copy' button to copy text to the clipboard, and then click the 'Paste' button to read back the text.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Copied text to clipboard using DataTransfer.
PASS finishedCopying became true
PASS Read text from clipboard using navigator.clipboard.readText.
PASS clipboardText is "garply"
PASS successfullyParsed is true

TEST COMPLETE
